+++ packages/contracts/test/utils/deployer.ts
@@ -9,7 +9,7 @@ import { web3 } from './web3_wrapper';
 const deployerOpts = {
     web3Provider: web3.currentProvider,
     artifactsDir: path.resolve('src', 'artifacts'),
-    jsonrpcPort: devConstants.RPC_PORT,
+    jsonrpcUrl: devConstants.RPC_URL,
     networkId: constants.TESTRPC_NETWORK_ID,
     defaults: {
         gas: devConstants.GAS_ESTIMATE,