+++ packages/deployer/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## v0.3.0 - _TBD, 2018_
 
     * Add support for Solidity 0.4.20 and 0.4.21
+    * Replace `jsonrpcPort` config by `jsonrpcUrl` (#426)
+    * Replace `jsonrpc-port` CLI option by `jsonrpc-url` (#426)
 
 ## v0.2.0 - _March 4, 2018_